backend/requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
Pillow==6.2.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
qrcode==6.1
